feat: 联调后完善代码
1.新旧InRelease不同时,移除所有packages;
2.增加系统更新工具错误码归类;
3.优化可更新内容判断;
4.优化重启后检查配置设置脚本;
遗留: cve增量数据处理，UUID缓存处理，离线系统环境检查

Log: 联调后完善代码
Influence: 开发中，暂不涉及
Task: https://pms.uniontech.com/task-view-290439.html

+++ var/lib/lastore/scripts/upgrade_check.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
-
-if [ -e "/etc/deepin/deepin_update_option.json" ]; then
+if [ -e "/etc/deepin/deepin_update_option.json" ] && [ ! -L "/tmp/deepin_update_option.json" ];  then
     # 如果文件存在，则创建软连接
     ln -s "/etc/deepin/deepin_update_option.json" "/tmp/deepin_update_option.json"
 else
